+++ server/dbi/sqlite.js
@@ -66,7 +66,7 @@ const SQ       = require('../libs/sqlite/api')
  */
 const SQLite = function(params) {
   this._name = 'sqlite';
-  this._db = params.database;
+  this._db = process.env.KAPP_TEST_MODE ? params.testdb : params.database;
   this._lib = SQ;
 };
 
@@ -78,44 +78,46 @@ const methods = {
   /**
    * Checks if the database is empty.
    *
-   * @method ()
+   * @method (arg1)
    * @public
-   * @param {}              -,
+   * @param {Object}        the connection to the db,
    * @returns {Boolean}     returns true if empty otherwise false,
    * @since 0.0.0
    */
-  async isDbEmpty() {
+  async isDbEmpty(cn) {
     const SQL = 'SELECT count(*) FROM sqlite_master WHERE type = "table"';
-    const resp = await this._lib.get(SQL);
+    const resp = await this._lib.get(cn, SQL);
     return resp['count(*)'] === 0;
   },
 
   /**
    * Checks if the table exists.
    *
-   * @method (arg1)
+   * @method (arg1, arg2)
    * @public
+   * @param {Object}        the connection to the db,
    * @param {String}        the name of the table,
    * @returns {Boolean}     returns true if the table exists otherwise false,
    * @since 0.0.0
    */
-  async isTable(table) {
+  async isTable(cn, table) {
     const sql = `SELECT name FROM sqlite_master WHERE type='table' AND name='${table}';`;
-    const r = await this._lib.get(sql);
+    const r = await this._lib.get(cn, sql);
     return typeof r === 'object' && r.name === table;
   },
 
   /**
    * Returns the table structure.
    *
-   * @method (arg1)
+   * @method (arg1, arg2)
    * @public
+   * @param {Object}        the connection to the db,
    * @param {String}        the name of the table,
    * @returns {Array}       returns the structure of the table,
    * @since 0.0.0
    */
-  async getTableStructure(table) {
-    const r = await this._lib.all(`pragma table_info(${table})`);
+  async getTableStructure(cn, table) {
+    const r = await this._lib.all(cn, `pragma table_info(${table})`);
     const struct = [];
 
     // Standardize:
@@ -148,29 +150,30 @@ const methods = {
   /**
    * Checks if the table is empty.
    *
-   * @method ()
+   * @method (arg1)
    * @public
-   * @param {}              -,
+   * @param {Object}        the connection to the db,
    * @returns {Boolean}     returns true if empty otherwise false,
    * @since 0.0.0
    */
-  async isTableEmpty(table) {
+  async isTableEmpty(cn, table) {
     const sql = `SELECT count(*) FROM ${table}`;
-    const r = await this._lib.get(sql);
+    const r = await this._lib.get(cn, sql);
     return r['count(*)'] === 0;
   },
 
   /**
    * Counts the number of tables, rows, columns.
    *
-   * @method (arg1, arg2)
+   * @method (arg1, arg2, arg3)
    * @public
+   * @param {Object}        the connection to the db,
    * @param {String}        what to count ('tables', 'rows', columns'),
    * @param {String}        the concerned table for rows or columns count,
    * @returns {Boolean}     returns true if empty otherwise false,
    * @since 0.0.0
    */
-  async count(what, table) {
+  async count(cn, what, table) {
     let sql
       , res
       ;
@@ -181,21 +184,21 @@ const methods = {
           SELECT count(*) FROM SQLITE_MASTER
             WHERE TYPE = 'table'
         `;
-        res = await this._lib.all(sql);
+        res = await this._lib.all(cn, sql);
         return res[0]['count(*)'] - 1;
 
       case 'columns':
         sql = `
           SELECT * FROM pragma_table_info('${table}')
         `;
-        res = await this._lib.all(sql);
+        res = await this._lib.all(cn, sql);
         return res.length;
 
       case 'rows':
         sql = `
         SELECT count(*) FROM ${table}
         `;
-        res = await this._lib.all(sql);
+        res = await this._lib.all(cn, sql);
         return res[0]['count(*)'];
 
       default: